design: 설정 화면 내 box간 섹션 분리 및 섹션명 덧붙임
